refactor: asset 이름 수정1
