+++ assets/icons/bigCheck.ts
